+++ .idea/deploymentTargetSelector.xml
@@ -19,6 +19,9 @@
       <SelectionState runConfigName="KorailTalkDropdownPreview">
         <option name="selectionMode" value="DROPDOWN" />
       </SelectionState>
+      <SelectionState runConfigName="CheckTrainCardPreview">
+        <option name="selectionMode" value="DROPDOWN" />
+      </SelectionState>
     </selectionStates>
   </component>
 </project>
